MTDSA-7078 Update Models for Retrieve SA tax calculation income tax NICs

+++ app/v1/models/response/getIncomeTaxAndNics/detail/GiftAid.scala
@@ -18,7 +18,9 @@ package v1.models.response.getIncomeTaxAndNics.detail
 
 import play.api.libs.json.{Json, OFormat}
 
-case class GiftAid(grossGiftAidPayments: BigDecimal, rate: BigDecimal, giftAidTax: BigDecimal)
+case class GiftAid(grossGiftAidPayments: BigInt,
+                   rate: BigDecimal,
+                   giftAidTax: BigDecimal)
 
 object GiftAid {
   implicit val format: OFormat[GiftAid] = Json.format[GiftAid]
